Organized Scripts into Directories and added to Include
Keeping things organized and tidy
