+++ Breakout/include/Breakout/Managers/GameManager.cpp
@@ -1,5 +1,5 @@
 #include "GameManager.h"
-#include "Ball.h"
+#include "Breakout/Ball/Ball.h"
 #include "PowerupManager.h"
 #include <iostream>
