+++ Breakout/include/Breakout/Paddle/Paddle.cpp
